kiwoomy/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
PyQt5==5.15.9
python-dateutil==2.8.2
pywin32==306; sys_platform == "win32" 